Move qltest-discovery.ts to /query-testing/qltest-discovery.ts
