+++ extensions/ql-vscode/src/query-testing/test-adapter.ts
@@ -18,7 +18,7 @@ import {
   QLTestNode,
   QLTestDirectory,
   QLTestDiscovery,
-} from "../qltest-discovery";
+} from "./qltest-discovery";
 import { Event, EventEmitter, CancellationTokenSource } from "vscode";
 import { DisposableObject } from "../pure/disposable-object";
 import { CodeQLCliServer, TestCompleted } from "../codeql-cli/cli";
